refactor the code on the landing page by moving most of the rendering code to its own components, and moving the api logic to a api folder
